+++ src/bcm2835.c
@@ -10,7 +10,7 @@ uint32_t getWordRegister(uint32_t addr){
     return *((volatile uint32_t*)addr);
 }
 
-uint8_t BCM2835_configGPIOPin(uint8_t pin, uint8_t funct){
+uint8_t BCM2835_configGPIOPin(uint8_t pin, uint16_t funct){ // TO DO ERROR WRITING same reg twice
 	uint32_t tmpReg;
 	uint8_t funSelRegOff = (pin / 10);
 	uint8_t funSelReg = funSelRegOff * 0x04U;
@@ -116,7 +116,7 @@ uint8_t BCM2835_receiveByteUART(){
 
  void __attribute__((optimize(0)))  BCM2835_soft_waitms(uint32_t milis){
 // void  BCM2835_soft_waitms(uint32_t milis){
-	//VALUE OBTAINED EXPERIMENTALLY
+	//CONSTANT VALUE OBTAINED EXPERIMENTALLY
 	uint32_t i, j;
 
 	for(j=milis; j > 0; j--){
